renamed evaluate.py added try catch clause to dependencies updated doc to link to readme file
