+++ docs/ExpertAgent.rst
@@ -5,7 +5,7 @@ Description
 -----------
 This Agent uses a greedy algorithm to try to solve an overflow by trying many different topological actions.
 
-To use this agent, please follow the install steps in ExpertAgent/READ_ME.md
+To use this agent, please follow the install steps in ExpertAgent/READ_ME.md : https://github.com/rte-france/l2rpn-baselines/tree/master/l2rpn_baselines/ExpertAgent/READ_ME.md
 
 You can find the documentation on the algorithm and it's usage on : https://github.com/marota/ExpertOp4Grid
