+++ l2rpn_baselines/ExpertAgent/__init__.py
@@ -5,6 +5,6 @@ __all__ = [
 ]
 
 from l2rpn_baselines.ExpertAgent.ExpertAgent import ExpertAgent
-from l2rpn_baselines.ExpertAgent.eval_expertagent import evaluate
+from l2rpn_baselines.ExpertAgent.evaluate import evaluate
 from l2rpn_baselines.ExpertAgent.ExpertAgent import other_rewards
